+++ src/lib/pages/proposal-details/components/proposal-overview/proposal-period-overview/VotingOverview.tsx
@@ -5,7 +5,7 @@ import { ErrorFetchingProposalInfos } from "../../ErrorFetchingProposalInfos";
 import { useInternalNavigate } from "lib/app-provider";
 import { CustomIcon } from "lib/components/icon";
 import { Loading } from "lib/components/Loading";
-import { TabIndex, PeriodIndex } from "lib/pages/proposal-details/types";
+import { TabIndex } from "lib/pages/proposal-details/types";
 import { ProposalStatus } from "lib/types";
 import { formatUTC } from "lib/utils";
 
@@ -68,7 +68,6 @@ const VotingOverviewBody = ({
             query: {
               id: proposalData.id,
               tab: TabIndex.Vote,
-              period: PeriodIndex.Voting,
             },
             options: {
               shallow: true,